requirements.txt
pygame~=1.9.6
numpy~=1.19.2
Pillow~=8.0.1
